Integrated authentication into Evy's RegistrationScreen design
