+++ src/screens/RegistrationScreen.css
@@ -12,6 +12,7 @@
     display: flex;
     flex-direction: column;
     background-color: #F7DEC0;
+    align-items: center;
 }
 
 .heading-text {
@@ -28,12 +29,14 @@
     display: flex;
     flex-direction: column;
     background-color: #E7C7B4;
+    align-items: center;
 }
 
 .prompt-style{
     font-size: 20px;
     font-weight: bold;
     color: #115D81;
+    margin-bottom: -10px;
 }
 
 .input-style{
@@ -42,7 +45,7 @@
     background-color: #2C95B5;
     border: none;
     border-radius: 10px;
-    margin: 10px;
+    margin-bottom: 40px;
 }
 
 button {
