+++ src/screens/RegistrationScreen.jsx
@@ -1,95 +1,89 @@
-import React, { useState } from 'react';
+import React, { useState } from "react";
 import './RegistrationScreen.css';
-import Login from '../components/login';
-import SignUp from '../components/signup';
+import {
+  createUserWithEmailAndPassword,
+  signInWithEmailAndPassword,
+} from "firebase/auth";
+import { auth } from "../firebase-config";
+
+const RegistrationScreen = ({ setShowProfileScreen }) => {
+  const [registerEmail, setRegisterEmail] = useState("");
+  const [registerPassword, setRegisterPassword] = useState("");
+  const [loginEmail, setLoginEmail] = useState("");
+  const [loginPassword, setLoginPassword] = useState("");
+
+  const register = async () => {
+    try {
+      const user = await createUserWithEmailAndPassword(
+        auth,
+        registerEmail,
+        registerPassword
+      );
+      console.log(user);
+      setShowProfileScreen(true); // Show the profile screen after successful registration
+    } catch (error) {
+      console.log(error.message);
+    }
+  };
+
+  const login = async () => {
+    try {
+      const user = await signInWithEmailAndPassword(
+        auth,
+        loginEmail,
+        loginPassword
+      );
+      console.log(user);
+      setShowProfileScreen(true); // Show the profile screen after successful login
+    } catch (error) {
+      console.log(error.message);
+    }
+  };
+
+  return (
+    <>
+    <div>
+        <div className='body-section'>
+            <div className = 'registration-section'>
+                <div className = 'heading-text'>Register</div>
+                <label class="prompt-style">Email:</label><br></br>
+                <input 
+                    class="input-style" 
+                    onChange={(event) => {
+                        setRegisterEmail(event.target.value);
+                      }}
+                    />
+                <label class="prompt-style">Password:</label><br></br>
+                <input 
+                    class="input-style" 
+                    onChange={(event) => {
+                        setRegisterPassword(event.target.value);
+                      }}
+                    />
+                    <button onClick={register}> Create User</button>
+            </div>
 
 
-const RegistrationScreen = () => { 
-
-const [username, setUsername] = useState('');
-const [password, setPassword] = useState('');
-const [email, setEmail] = useState('');
-
-const handleRegister =() => {
-    console.log({username, email, password});
-}
-
-const [loginUsername, setLoginUsername] = useState('');
-const [loginPassword, setLoginPassword] = useState('');
-
-const handleLogin =() => {
-    console.log({loginUsername, loginPassword});
-}
-
-
-return (
-<div>
-    <body>
-    <section className='body-section'>
-
-
-        <section className = 'registration-section'>
-            <section className = 'heading-text'>Register</section>
-            <form align="center">
-            <label for="username" title="username" class="prompt-style">Username:</label><br></br>
-            <input 
-                class="input-style" 
-                type="text" 
-                id="username" 
-                value={username}
-                onChange={(e) => setUsername(e.target.value)}/>
-                <br></br>
-            <label for="email" title="email" class="prompt-style">Email:</label><br></br>
-            <input 
-                class="input-style" 
-                type="email" 
-                id="email" 
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                />
-                <br></br>
-            <label for="password" title="password" class="prompt-style">Password:</label><br></br>
-            <input 
-                class="input-style" 
-                type="text" 
-                id="password" 
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}/>
-                <br></br>
-                <button variant="primary" onclick={handleRegister} >Register</button>
-            </form>
-        </section>
-
-
-        <section className = 'login-section'>
-            <section className = 'heading-text'>Login</section>
-            <form align="center">
-            <label for="loginUsername" title="loginUsername" class="prompt-style">Username:</label><br></br>
-            <input 
-                class="input-style" 
-                type="text" 
-                id="loginUsername" 
-                value={loginUsername}
-                onChange={(e) => setLoginUsername(e.target.value)}/>
-                <br></br>
-            <label for="loginPassword" title="loginPassword" class="prompt-style">Password:</label><br></br>
-            <input 
-                class="input-style" 
-                type="text" 
-                id="loginPassword" 
-                value={loginPassword}
-                onChange={(e) => setLoginPassword(e.target.value)}/>
-                <br></br>
-                <button variant="primary" onclick={handleLogin} >Login</button>
-            </form>
-        </section>
-
-        
-    </section>
-    </body>
-</div>
-);
-
+            <div className = 'login-section'>
+                <div className = 'heading-text'>Login</div>
+                <label class="prompt-style">Email:</label><br></br>
+                <input 
+                    class="input-style" 
+                    onChange={(event) => {
+                        setLoginEmail(event.target.value);
+                      }}/>
+                <label class="prompt-style">Password:</label><br></br>
+                <input 
+                    class="input-style" 
+                    onChange={(event) => {
+                        setLoginPassword(event.target.value);
+                      }}/>
+                    <button onClick={login}> Login</button>
+            </div>
+        </div>
+    </div>
+    </>
+  );
 };
 
 export default RegistrationScreen;
